refactor(LambdaClient): 타임아웃 시간 자료형 int -> long으로 변경

+++ mapshot-image/src/main/java/com/joebrooks/mapshot/client/LambdaClient.java
@@ -18,7 +18,7 @@ public class LambdaClient extends CommonClient {
     private String path;
 
     public List<ImageResponse> sendRequest(ImageRequest request) {
-        int timeoutMillis = 30 * 1000;
+        long timeoutMillis = 30 * 1000L;
 
         String url = UriComponentsBuilder.newInstance()
                 .scheme("https")
